Add error handling for node name lookup in nvtop script

+++ Container-Root/hyperpod/ops/eks/nvtop.sh
@@ -17,6 +17,10 @@ if [ "$1" == "" ]; then
 else
 	node_name=$1
 	full_node_name=$(kubectl get nodes | grep $node_name | head -n 1 | cut -d ' ' -f 1)
+	if [ -z "$full_node_name" ]; then
+		echo "ERROR: no node matches '$node_name'" >&2
+		exit 1
+	fi
 	host_name=$(echo $full_node_name | cut -d '.' -f 1)
 	pod_name=nvtop-${host_name}
 	CMD="kubectl run -it --rm $pod_name --image iankoulski/do-nvtop:latest --overrides='{\"apiVersion\": \"v1\", \"spec\": {\"nodeSelector\": { \"kubernetes.io/hostname\": \"$full_node_name\" }}}' --command -- nvtop"
